# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Catering Webapp API",
        "description": "Dokumentasi API untuk sistem pemesanan paket katering (role Penjual & Pelanggan).",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/penjual/lauks": {
            "get": {
                "tags": [
                    "Penjual - Lauk"
                ],
                "summary": "Daftar semua lauk",
                "operationId": "941d3e94bc8edd10c7bad5f942b3530f",
                "responses": {
                    "200": {
                        "description": "Daftar lauk (paginated)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Penjual - Lauk"
                ],
                "summary": "Tambah lauk baru",
                "operationId": "31f1ff5d242118051bcc9113ac7f645e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_lauk"
                                ],
                                "properties": {
                                    "nama_lauk": {
                                        "type": "string",
                                        "example": "Ayam Bakar"
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status_aktif": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lauk berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/penjual/lauks/{lauk}": {
            "get": {
                "tags": [
                    "Penjual - Lauk"
                ],
                "summary": "Detail 1 lauk",
                "operationId": "c7cbcefb9e1f704b1c9d39f4c9287d8e",
                "parameters": [
                    {
                        "name": "lauk",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail lauk"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Penjual - Lauk"
                ],
                "summary": "Update lauk",
                "operationId": "bc3501c6a0343447ae9ea7690038993c",
                "parameters": [
                    {
                        "name": "lauk",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_lauk"
                                ],
                                "properties": {
                                    "nama_lauk": {
                                        "type": "string"
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status_aktif": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lauk berhasil diupdate"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Penjual - Lauk"
                ],
                "summary": "Hapus lauk",
                "description": "Ditolak (422) kalau lauk ini sudah pernah dipilih di pesanan.",
                "operationId": "950df8c7373393ec47c22be69054e433",
                "parameters": [
                    {
                        "name": "lauk",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lauk berhasil dihapus"
                    },
                    "422": {
                        "description": "Sudah pernah dipilih di pesanan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/penjual/pesanan": {
            "get": {
                "tags": [
                    "Penjual - Pesanan"
                ],
                "summary": "Daftar semua pesanan (bisa difilter status)",
                "operationId": "fd252ff24c1d74d3e8fc82caf82589e6",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter berdasarkan status_pesanan",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "menunggu_validasi",
                                "disetujui",
                                "ditolak",
                                "batal",
                                "selesai"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar pesanan (paginated)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/penjual/pesanan/{pesanan}/validasi": {
            "patch": {
                "tags": [
                    "Penjual - Pesanan"
                ],
                "summary": "Setujui atau tolak pesanan yang masuk",
                "operationId": "7ddfe09cd077b54b890f521692bc65aa",
                "parameters": [
                    {
                        "name": "pesanan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status_pesanan"
                                ],
                                "properties": {
                                    "status_pesanan": {
                                        "type": "string",
                                        "enum": [
                                            "disetujui",
                                            "ditolak"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status berhasil diupdate"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/penjual/pesanan/{pesanan}/produksi": {
            "patch": {
                "tags": [
                    "Penjual - Pesanan"
                ],
                "summary": "Update status dapur/produksi pesanan",
                "operationId": "ed8d5a2800a0640f9664b7549cdc9ea7",
                "parameters": [
                    {
                        "name": "pesanan",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status_produksi"
                                ],
                                "properties": {
                                    "status_produksi": {
                                        "type": "string",
                                        "enum": [
                                            "belum_diproses",
                                            "diproses",
                                            "selesai"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status berhasil diupdate"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login dulu lewat /login, lalu isi token yang didapat di sini (tanpa kata \"Bearer\").",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}